pkg/agent: make Run return error
This allows to track when agent finishes it's run and inspecting it's
results, which is useful in testing.

This commit adds also a basic test inspecting the run error, for now
without a positive case where no error is returned, as we have no way of
mocking/configuring certain dependencies, for example customizing paths
of Flatcar configuration.

Closes #113

Signed-off-by: Mateusz Gozdek <mgozdek@microsoft.com>